refactor(models): land H1 programme shapes in models/h1.py, not the shared dump
models.py mixed two concerns: HackerOne-specific catalog/scope types
(ProgrammePreview, Programme, ScopeItem, ScopeType) and the cross-agent
artefact types every pipeline stage exchanges (Endpoint, ReconResult,
RawFinding, ...). The H1 surface is the part most likely to grow as the
client picks up more endpoints; keeping it in the shared dump invites the
same "split the file" smell pylint enforces at 500 lines (max-module-lines),
caught earlier by the concerns mixing rather than the line count.

Convert models.py to a package (models/__init__.py) and move the four H1
types into models/h1.py. ReconResult.programme keeps its typed reference by
importing Programme back into the package init after Severity is defined (h1
depends on Severity, so the import sits below the enum to break the cycle
cleanly). No back-compat re-export from the package root - every importer
touching an H1 name on this branch migrates to "from models.h1 import ..."
directly. Future per-domain splits (recon, findings, ...) are deliberately
out of scope.

Mirrors the models/attack.py split landing in parallel on #137.
