+++ squad/vulnerability_researcher/__init__.py
@@ -6,7 +6,8 @@ from pathlib import Path
 
 from crewai.tools import tool
 
-from models import Programme, Severity
+from models import Severity
+from models.h1 import Programme
 from squad import SquadMember, read_run_file_tool, read_run_filelist_tool
 from tools import cwe_data, http, owasp_data
 from tools.h1_api import h1
