+++ tools/pentest/triage.py
@@ -13,7 +13,8 @@ from __future__ import annotations
 import logging
 
 from config import config
-from models import Programme, RawFinding
+from models import RawFinding
+from models.h1 import Programme
 from tools import http
 from tools.recon.scope import extract_domain, filter_in_scope
